# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/usuarios/":{"get":{"tags":["Usuários"],"summary":"Listar Usuarios","operationId":"listar_usuarios_api_usuarios__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/operadoras":{"get":{"summary":"Lista Operadoras","operationId":"lista_operadoras_api_operadoras_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OperadoraCadastral"},"title":"Response Lista Operadoras Api Operadoras Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/operadoras/{cnpj}":{"get":{"summary":"Detalhar Operadora","operationId":"detalhar_operadora_api_operadoras__cnpj__get","parameters":[{"name":"cnpj","in":"path","required":true,"schema":{"type":"string","title":"Cnpj"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperadoraCadastral"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/operadoras/{cnpj}/despesas":{"get":{"summary":"Listar Despesas Operadora","operationId":"listar_despesas_operadora_api_operadoras__cnpj__despesas_get","parameters":[{"name":"cnpj","in":"path","required":true,"schema":{"type":"string","title":"Cnpj"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DespesasConsolidadas"},"title":"Response Listar Despesas Operadora Api Operadoras  Cnpj  Despesas Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/estatisticas":{"get":{"summary":"Obter Estatisticas","operationId":"obter_estatisticas_api_estatisticas_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/estatisticas/despesas-por-uf":{"get":{"summary":"Despesas Por Uf","operationId":"despesas_por_uf_api_estatisticas_despesas_por_uf_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"DespesasConsolidadas":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"cnpj":{"type":"string","title":"Cnpj"},"razao_social":{"type":"string","title":"Razao Social"},"ano":{"type":"integer","title":"Ano"},"trimestre":{"type":"integer","title":"Trimestre"},"valor_despesa":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*(?:\\d{0,18}|(?=[\\d.]{1,21}0*$)\\d{0,18}\\.\\d{0,2}0*$)","title":"Valor Despesa"}},"type":"object","required":["cnpj","razao_social","ano","trimestre","valor_despesa"],"title":"DespesasConsolidadas"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OperadoraCadastral":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"cnpj":{"type":"string","title":"Cnpj"},"razao_social":{"type":"string","title":"Razao Social"},"ano":{"type":"integer","title":"Ano"},"trimestre":{"type":"integer","title":"Trimestre"},"cnpj_valido":{"type":"boolean","title":"Cnpj Valido","default":false},"registro_ans":{"type":"string","title":"Registro Ans"},"modalidade":{"type":"string","title":"Modalidade"},"uf":{"type":"string","title":"Uf"}},"type":"object","required":["cnpj","razao_social","ano","trimestre","registro_ans","modalidade","uf"],"title":"OperadoraCadastral"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}